+++ hunmin-application/src/main/kotlin/com/application/auth/AuthService.kt
@@ -1,15 +1,15 @@
 package com.application.auth
 
-import com.application.auth.port.`in`.AuthUseCase
-import com.application.auth.port.`in`.command.SignInCommand
-import com.application.auth.port.`in`.command.SignUpCommand
-import com.application.auth.port.out.AuthRepositoryPort
-import com.application.auth.port.out.TokenProviderPort
+import com.domain.auth.port.`in`.AuthUseCase
+import com.domain.auth.port.`in`.command.SignInCommand
+import com.domain.auth.port.`in`.command.SignUpCommand
+import com.domain.auth.port.out.AuthRepositoryPort
+import com.domain.auth.port.out.TokenProviderPort
 import com.common.global.exceptions.base.CustomException
 import com.common.util.throwWhen
 import com.domain.auth.Auth
 import com.domain.auth.exception.AuthExceptionType
-import com.domain.auth.service.AuthPasswordEncryptorPort
+import com.domain.auth.port.out.AuthPasswordEncryptorPort
 import org.springframework.stereotype.Service
 
 @Service
