+++ hunmin-application/src/testFixtures/kotlin/com/application/auth/AuthCommandFixture.kt
@@ -1,7 +1,7 @@
 package com.application.auth
 
-import com.application.auth.port.`in`.command.SignInCommand
-import com.application.auth.port.`in`.command.SignUpCommand
+import com.domain.auth.port.`in`.command.SignInCommand
+import com.domain.auth.port.`in`.command.SignUpCommand
 import com.domain.auth.Auth
 
 class AuthCommandFixture {
