+++ hunmin-domain/src/main/kotlin/com/domain/auth/port/in/AuthUseCase.kt
@@ -1,0 +1,11 @@
+package com.domain.auth.port.`in`
+
+import com.domain.auth.port.`in`.command.SignInCommand
+import com.domain.auth.port.`in`.command.SignUpCommand
+
+interface AuthUseCase {
+
+    fun signUp(command: SignUpCommand): String
+
+    fun signIn(command: SignInCommand): String
+}
