+++ hunmin-domain/src/test/kotlin/com/domain/auth/service/StubAuthPasswordEncryptorAdaptor.kt
@@ -1,5 +1,7 @@
 package com.domain.auth.service
 
+import com.domain.auth.port.out.AuthPasswordEncryptorPort
+
 class StubAuthPasswordEncryptorAdaptor : AuthPasswordEncryptorPort {
 
     override fun encrypt(password: String): String {
